Changed parameter of min function from double to min and reverted back to original round2 solution (keeping round2c arround)

+++ Calculate Library/src/Calculate.java
@@ -111,7 +111,7 @@ public class Calculate {
     	return max;
     }
     
-    public static double min(double a , double b ) {
+    public static int min(int a , int b ) {
         if (a < b) 
         { 
              return a;
